build: improve webpack config
