+++ apps/core/src/index.tsx
@@ -3,7 +3,6 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
 async function main() {
-  await import('./bootstrap/before-app');
   const { App } = await import('./app');
   const root = document.getElementById('app');
   assertExists(root);
